Feature/songdetail
확인했어요

+++ static/script/song_detail.js
@@ -73,8 +73,8 @@ async function VoiceDetailView(id){
     response_json = await response.json()
 
     if (response.status == 200) {
-        let VoicelInfo = response_json
-        return VoiceInfo
+        let VoiceDetailInfo = response_json
+        return VoiceDetailInfo
 
     }else {
         alert(response_json["error"])
@@ -114,7 +114,8 @@ async function commentView(id, content){
 
 //댓글 Delete
 async function deletecommentView(id, comment_id){
-    alert("댓글이 삭제되었습니다.")
+    var delConfirm = confirm("정말 파일을 삭제하시겠습니까?")
+        if(delConfirm){
         const response = await fetch(`${backendBaseUrl}/songs/${id}/comment/${comment_id}`, {
             method: 'DELETE',
             headers: {
@@ -125,11 +126,12 @@ async function deletecommentView(id, comment_id){
         })
         response_json = await response.json
         if (response.status === 200) {
+            alert("댓글이 삭제되었습니다.")
             return response_json
         }else {
             alert(response_json["error"])
         }
-}
+}}
 
 //댓글 Update
 async function updatecommentView(id, comment_id, comment){
@@ -161,8 +163,8 @@ async function updatecommentView(id, comment_id, comment){
 
 //좋아요 POST
 async function SongLike() {
-
     const response = await fetch(`${backendBaseUrl}/songs/${likeId}/song_like/`, {
+
         method: 'POST',
         headers: {
             Accept:"application/json",
@@ -171,9 +173,11 @@ async function SongLike() {
         },
     }
     )
+
     const a_heart_btn = document.getElementById("heartBtn")
 
     response_json = await response.json
+
     if (response.status == 200) {
         window.location.reload()
         return
@@ -181,8 +185,52 @@ async function SongLike() {
         alert(response_json["msg"])
     }
     
+
 }
 
+
+async function mpload() {
+    let voice = document.querySelector('#mp-file-upload');
+    let voice_file = voice.files[0]
+    formData = new FormData();
+    
+    formData.append("recode",voice_file)
+    console.log(formData)
+    const response = await fetch(`${backendBaseUrl}/songs/${songId}/voice/`, {
+        method: "POST",
+        headers: {
+        "Authorization": "Bearer " + localStorage.getItem("access")
+    },
+    body: formData
+    })
+    if (response.status == 201) {
+        alert('저장됐습니다.')
+        window.location.reload()
+    }else {
+        alert(response_json["msg"])
+    }
+}
+
+async function deletevoiceView(id, voice_id){
+    var delConfirm = confirm("정말 파일을 삭제하시겠습니까?")
+        if (delConfirm){
+        const response = await fetch(`${backendBaseUrl}/songs/${id}/voice/${voice_id}`, {
+            method: 'DELETE',
+            headers: {
+                Accept: "application/json",
+                "Content-type": "application/json",
+                "Authorization": "Bearer " + localStorage.getItem("access")
+            }
+        })
+        response_json = await response.json
+        if (response.status === 200) {
+            alert("파일이 삭제되었습니다.")
+        }else {
+            alert(response_json["error"])
+        }
+}
+
+}
 
 // 좋아요 불러오기
 async function songlistview2() {
